Change catalogid and micrositeid ties

+++ lib/tasks/compile_microsite.rake
@@ -1,4 +1,7 @@
 task compile_microsite: :environment do
+  if !Rails.env.production?
+    raise 'Re-run with "RAILS_ENV=production"'
+  end
   Dir.mkdir(Rails.root.join('compiled')) unless Dir.exist? Rails.root.join('compiled')
   `rm -rf compiled/*`
   `rm compiled.zip`
@@ -20,6 +23,7 @@ task compile_microsite: :environment do
   body = session.response.body
   body.sub!(/<link[^>]*>/, '{% for style_url in style_urls %} <link type="text/css" href="{{style_url}}" media="all" rel="stylesheet"/> {% endfor %}')
   body.sub!(/<script[^>]*>[^<]*<\/script>/, ' {% for script_url in script_urls %} <script type="text/javascript" src="{{script_url}}"> </script> {% endfor %}')
+  body.sub!(/catalog\_tie/, '<script type="text/javascript">window.catalog_id={{catalog_id}};window.microsite_id={{microsite_id}}</script>')
   File.open(Rails.root.join('compiled', 'index.html.liquid'), 'w') do |f|
     f << body
   end
